cloud_run_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
ultralytics==8.3.55
google-cloud-storage==2.13.0
python-dotenv==1.0.0
torch==2.0.1
opencv-python-headless==4.8.1.78